+++ grails-datastore-gorm/src/main/groovy/org/grails/datastore/gorm/validation/jakarta/ConfigurableParameterNameProvider.groovy
@@ -1,4 +1,4 @@
-package org.grails.datastore.gorm.validation.javax
+package org.grails.datastore.gorm.validation.jakarta
 
 import groovy.transform.CompileStatic
