+++ grails-datastore-gorm/src/main/groovy/org/grails/datastore/gorm/validation/jakarta/MethodKey.java
@@ -1,4 +1,4 @@
-package org.grails.datastore.gorm.validation.javax;
+package org.grails.datastore.gorm.validation.jakarta;
 
 import java.util.Arrays;
