+++ grails-datastore-gorm/src/main/groovy/org/grails/datastore/gorm/validation/jakarta/services/implementers/MethodValidationImplementer.groovy
@@ -1,4 +1,4 @@
-package org.grails.datastore.gorm.validation.javax.services.implementers
+package org.grails.datastore.gorm.validation.jakarta.services.implementers
 
 import groovy.transform.CompileStatic
 import org.codehaus.groovy.ast.*
@@ -7,8 +7,8 @@ import org.codehaus.groovy.ast.stmt.BlockStatement
 import org.codehaus.groovy.ast.stmt.Statement
 import org.grails.datastore.gorm.services.ServiceEnhancer
 import org.grails.datastore.gorm.transform.AbstractTraitApplyingGormASTTransformation
-import org.grails.datastore.gorm.validation.javax.ConfigurableParameterNameProvider
-import org.grails.datastore.gorm.validation.javax.services.ValidatedService
+import org.grails.datastore.gorm.validation.jakarta.ConfigurableParameterNameProvider
+import org.grails.datastore.gorm.validation.jakarta.services.ValidatedService
 import org.grails.datastore.mapping.reflect.ClassUtils
 
 import jakarta.validation.Constraint
