+++ grails-datastore-gorm/src/main/groovy/org/grails/datastore/gorm/validation/registry/support/ValidatorRegistries.java
@@ -1,7 +1,7 @@
 package org.grails.datastore.gorm.validation.registry.support;
 
 import org.grails.datastore.gorm.validation.constraints.registry.DefaultValidatorRegistry;
-import org.grails.datastore.gorm.validation.javax.JavaxValidatorRegistry;
+import org.grails.datastore.gorm.validation.jakarta.JakartaValidatorRegistry;
 import org.grails.datastore.mapping.core.connections.ConnectionSourceSettings;
 import org.grails.datastore.mapping.model.MappingContext;
 import org.grails.datastore.mapping.reflect.ClassUtils;
@@ -42,7 +42,7 @@ public class ValidatorRegistries {
     public static ValidatorRegistry createValidatorRegistry(MappingContext mappingContext, ConnectionSourceSettings settings, MessageSource messageSource ) {
         ValidatorRegistry validatorRegistry;
         if(isJavaxValidationAvailable()) {
-            validatorRegistry = new JavaxValidatorRegistry(mappingContext, settings, messageSource);
+            validatorRegistry = new JakartaValidatorRegistry(mappingContext, settings, messageSource);
         }
         else {
             validatorRegistry = new DefaultValidatorRegistry(mappingContext, settings, messageSource);
